requirements.txt
groq
langchain
faiss-cpu
PyPDF2
pdfplumber
dotenv